separate out url, remove promises
