+++ karma.conf.js
@@ -39,7 +39,8 @@ module.exports = function(config) {
 
     [!options.ie8 ? (!options['babel'] ? 'node_modules/traceur/bin/traceur.js' : 'node_modules/babel-core/browser.js') : ''],
 
-    'dist/es6-module-loader' + (options.polyfill ? '' : '-sans-promises') + '.src.js',
+    [options.polyfill ? 'node_modules/when/es6-shim/Promise' : ''],
+    'dist/es6-module-loader.src.js',
 
     'test/_browser.js',
     'test/browser-script-type-module.js',
